Abstract the FetchBLobstoreUrl and sendFormData functions, so that they can be used to add a picture to a user profile and listing object.
