+++ src/main/webapp/newlisting.html
@@ -133,8 +133,7 @@
             </form>
           </li>
         </ul>
-        <a href="user.html">"Submit"</a>
-        <button>Submit</button>
+        <button type="button">Submit</button>
         <button type="button" onclick="displayPreviewListing()">
           Make Preview
         </button>
